feat: re-run checks if base branch has modified config file
