+++ docs/changelog.rst
@@ -1,8 +1,9 @@
 CHANGELOG
 =====================================
 
+| July 1, 2020 : When config file is added/modified in base branch, mergeable will trigger for all PR against the base branch `#153 <https://github.com/mergeability/mergeable/issues/153>`_
 | June 5, 2020 : For missing fields in 'checks', default values will be used `#233 <https://github.com/mergeability/mergeable/issues/233#issuecomment-632211789>`_
-| May 30, 2020 : New Action `merge` added `#201 <https://github.com/mergeability/mergeable/issues/228>`_
+| May 30, 2020 : New Action `merge` added `#201 <https://github.com/mergeability/mergeable/issues/201>`_
 | May 29, 2020 : throw `UnSupportedSettingError` if provided setting is not valid `#228 <https://github.com/mergeability/mergeable/issues/228>`_
 | May 29, 2020 : Ability to Limit `stale` validator to certain days and time `#221 <https://github.com/mergeability/mergeable/issues/221>`_
 | May 23, 2020 : Allow PRs/Issues to be assigned to their author by using `@author` in the `assign` action
