+++ lib/configuration/configuration.js
@@ -113,6 +113,7 @@ class Configuration {
 
     // probotContext.config loads config from current repo or from a repo called
     // '.github' in the same organisation as a fallback. It returns the parsed YAML object.
+
     const configPath = process.env.CONFIG_PATH ? process.env.CONFIG_PATH : 'mergeable.yml'
     const config = await context.probotContext.config(configPath)
     if (config === null || config === undefined) {
