+++ lib/flex/lib/createPromises.js
@@ -1,0 +1,15 @@
+const createPromises = (arrayToIterate, registryName, funcCall, context, registry, name, result) => {
+  let promises = []
+  arrayToIterate.forEach(element => {
+    let key = element.do
+
+    let klass = registry[registryName].get(key)
+    let eventName = `${context.event}.${context.payload.action}`
+    if (klass.isEventSupported(eventName)) {
+      promises.push(funcCall(klass, context, element, name, result))
+    }
+  })
+  return promises
+}
+
+module.exports = createPromises
