+++ lib/interceptors/index.js
@@ -1,6 +1,7 @@
 const REGISTRY = [
   new (require('./checkReRun'))(),
-  new (require('./milestoned'))()
+  new (require('./milestoned'))(),
+  new (require('./push'))()
 ] // eslint-disable-line
 
 /**
