+++ lib/mergeable.js
@@ -1,5 +1,5 @@
 const createScheduler = require('probot-scheduler')
-const flexExecutor = require('./flex')
+const flexExecutor = require('./flex/flex')
 const Context = require('./context')
 const logger = require('./logger')
 const _ = require('lodash')
