+++ lib/validators/approvals.js
@@ -13,18 +13,8 @@ class Approvals extends Validator {
   constructor () {
     super('approvals')
     this.supportedEvents = [
-      'pull_request.opened',
-      'pull_request.edited',
-      'pull_request_review.submitted',
-      'pull_request_review.edited',
-      'pull_request_review.dismissed',
-      'pull_request.labeled',
-      'pull_request.milestoned',
-      'pull_request.demilestoned',
-      'pull_request.assigned',
-      'pull_request.unassigned',
-      'pull_request.unlabeled',
-      'pull_request.synchronize'
+      'pull_request.*',
+      'pull_request_review.*'
     ]
 
     this.supportedSettings = {
