+++ lib/validators/assignee.js
@@ -4,18 +4,8 @@ class Assignee extends Validator {
   constructor () {
     super('assignee')
     this.supportedEvents = [
-      'pull_request.opened',
-      'pull_request.edited',
-      'pull_request_review.submitted',
-      'pull_request_review.edited',
-      'pull_request_review.dismissed',
-      'pull_request.labeled',
-      'pull_request.milestoned',
-      'pull_request.demilestoned',
-      'pull_request.unlabeled',
-      'pull_request.assigned',
-      'pull_request.unassigned',
-      'pull_request.synchronize',
+      'pull_request.*',
+      'pull_request_review.*',
       'issues.*'
     ]
