+++ lib/validators/lib/andOrValidatorProcessor.js
@@ -1,5 +1,5 @@
 const Register = require('../../register')
-const { getValidatorPromises } = require('../../flex')
+const getValidatorPromises = require('../../flex/lib/getValidatorPromises')
 const consolidateResult = require('../options_processor/options/lib/consolidateResults')
 const constructErrorOutput = require('../options_processor/options/lib/constructErrorOutput')
