+++ lib/validators/title.js
@@ -4,17 +4,8 @@ class Title extends Validator {
   constructor () {
     super('title')
     this.supportedEvents = [ 'pull_request.opened',
-      'pull_request.edited',
-      'pull_request_review.submitted',
-      'pull_request_review.edited',
-      'pull_request_review.dismissed',
-      'pull_request.labeled',
-      'pull_request.milestoned',
-      'pull_request.demilestoned',
-      'pull_request.assigned',
-      'pull_request.unassigned',
-      'pull_request.unlabeled',
-      'pull_request.synchronize',
+      'pull_request.*',
+      'pull_request_review.*',
       'issues.*'
     ]
     this.supportedSettings = {
